Port to Python3 for the root package (but not the wheel).
This patch modifies the build system and some parts of the code
to handle both toolchains with Python2 and toolchains with Python3,
with a preference for Python3 if both exist.

It only ports the root package, and not the wheel generation code.

Reviewed-on: http://gerrit2.aldebaran.lan/171
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>

+++ src/pyobjectconverter.cpp
@@ -314,7 +314,7 @@ class PyObjectLongTypeInterface: public PyObjectTypeInterface<qi::IntTypeInterfa
 {
 public:
   PYTYPE_GETSET(int64_t, LongLong, Long)
-  virtual unsigned int size() { return sizeof(long); }
+  virtual unsigned int size() { return sizeof(long long); }
   virtual bool isSigned() { return true; }
   PYTYPE_GETTYPE(PyObjectLongTypeInterface)
 };
